+++ lib/Latte/Nodes/ScriptNode.php
@@ -5,4 +5,5 @@ namespace JanHerman\Barista\Latte\Nodes;
 class ScriptNode extends SfcNode
 {
     protected const Languages = ['js', 'ts'];
+    protected const Type = 'script';
 }