+++ lib/Latte/Nodes/StyleNode.php
@@ -5,4 +5,5 @@ namespace JanHerman\Barista\Latte\Nodes;
 class StyleNode extends SfcNode
 {
     protected const Languages = ['css', 'scss'];
+    protected const Type = 'style';
 }